[Pal] Fix return type in INTERNAL_SYSCALL macros

+++ Pal/include/arch/x86_64/Linux/sysdep-arch.h
@@ -130,14 +130,14 @@
 
 #define INTERNAL_SYSCALL_NCS(name, err, nr, args...)     \
     ({                                                   \
-        unsigned long resultvar;                         \
+        long resultvar;                                  \
         LOAD_ARGS_##nr(args);                            \
         LOAD_REGS_##nr;                                  \
         __asm__ volatile(DO_SYSCALL                      \
                          : "=a"(resultvar)               \
                          : "0"(name)ASM_ARGS_##nr        \
                          : "memory", "cc", "r11", "cx"); \
-        (long)resultvar;                                 \
+        resultvar;                                       \
     })
 
 #undef INTERNAL_SYSCALL
@@ -162,51 +162,51 @@
 #define LOAD_REGS_0
 #define ASM_ARGS_0
 
-#define LOAD_ARGS_1(a1)           \
-    long int __arg1 = (long)(a1); \
+#define LOAD_ARGS_1(a1)       \
+    long __arg1 = (long)(a1); \
     LOAD_ARGS_0()
-#define LOAD_REGS_1                                \
-    register long int _a1 __asm__("rdi") = __arg1; \
+#define LOAD_REGS_1                            \
+    register long _a1 __asm__("rdi") = __arg1; \
     LOAD_REGS_0
 #define ASM_ARGS_1 ASM_ARGS_0, "r"(_a1)
 
-#define LOAD_ARGS_2(a1, a2)       \
-    long int __arg2 = (long)(a2); \
+#define LOAD_ARGS_2(a1, a2)   \
+    long __arg2 = (long)(a2); \
     LOAD_ARGS_1(a1)
-#define LOAD_REGS_2                                \
-    register long int _a2 __asm__("rsi") = __arg2; \
+#define LOAD_REGS_2                            \
+    register long _a2 __asm__("rsi") = __arg2; \
     LOAD_REGS_1
 #define ASM_ARGS_2 ASM_ARGS_1, "r"(_a2)
 
-#define LOAD_ARGS_3(a1, a2, a3)   \
-    long int __arg3 = (long)(a3); \
+#define LOAD_ARGS_3(a1, a2, a3) \
+    long __arg3 = (long)(a3);   \
     LOAD_ARGS_2(a1, a2)
-#define LOAD_REGS_3                                \
-    register long int _a3 __asm__("rdx") = __arg3; \
+#define LOAD_REGS_3                            \
+    register long _a3 __asm__("rdx") = __arg3; \
     LOAD_REGS_2
 #define ASM_ARGS_3 ASM_ARGS_2, "r"(_a3)
 
 #define LOAD_ARGS_4(a1, a2, a3, a4) \
-    long int __arg4 = (long)(a4);   \
+    long __arg4 = (long)(a4);       \
     LOAD_ARGS_3(a1, a2, a3)
-#define LOAD_REGS_4                                \
-    register long int _a4 __asm__("r10") = __arg4; \
+#define LOAD_REGS_4                            \
+    register long _a4 __asm__("r10") = __arg4; \
     LOAD_REGS_3
 #define ASM_ARGS_4 ASM_ARGS_3, "r"(_a4)
 
 #define LOAD_ARGS_5(a1, a2, a3, a4, a5) \
-    long int __arg5 = (long)(a5);       \
+    long __arg5 = (long)(a5);           \
     LOAD_ARGS_4(a1, a2, a3, a4)
-#define LOAD_REGS_5                               \
-    register long int _a5 __asm__("r8") = __arg5; \
+#define LOAD_REGS_5                           \
+    register long _a5 __asm__("r8") = __arg5; \
     LOAD_REGS_4
 #define ASM_ARGS_5 ASM_ARGS_4, "r"(_a5)
 
 #define LOAD_ARGS_6(a1, a2, a3, a4, a5, a6) \
-    long int __arg6 = (long)(a6);           \
+    long __arg6 = (long)(a6);               \
     LOAD_ARGS_5(a1, a2, a3, a4, a5)
-#define LOAD_REGS_6                               \
-    register long int _a6 __asm__("r9") = __arg6; \
+#define LOAD_REGS_6                           \
+    register long _a6 __asm__("r9") = __arg6; \
     LOAD_REGS_5
 #define ASM_ARGS_6 ASM_ARGS_5, "r"(_a6)
